+++ fixtures/application/sample2.tsx
@@ -1,0 +1,2 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<tileset version="1.8" tiledversion="1.8.0" name="test" tilewidth="1" tileheight="1" tilecount="0" columns="0"/>