+++ src/readers.rs
@@ -445,6 +445,10 @@ impl crate::FileFormat {
                 return Ok(Self::TrainingCenterXml);
             } else if contains(&buffer, b"<tt xmlns=\"http://www.w3.org/ns/ttml\"") {
                 return Ok(Self::TimedTextMarkupLanguage);
+            } else if contains(&buffer, b"<map") {
+                return Ok(Self::TiledMapXml);
+            } else if contains(&buffer, b"<tileset") {
+                return Ok(Self::TiledTilesetXml);
             } else if contains(&buffer, b"<USFSubtitles") {
                 return Ok(Self::UniversalSubtitleFormat);
             } else if contains(&buffer, b"<xliff") {